investafrik/requirements.txt
Django==5.0.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.2.0
psycopg2-binary==2.9.11
Pillow==10.4.0
python-decouple==3.8
django-extensions==3.2.3
django-debug-toolbar==4.4.6
djangorestframework-simplejwt==5.3.0
django-filter==24.2
django-ratelimit==4.1.0
coverage==7.6.1
factory-boy==3.3.0
requests==2.32.3
redis==5.0.8
celery==5.3.4
django-ckeditor==6.7.1
django-taggit==5.0.1
django-crispy-forms==2.3
crispy-tailwind==0.5.0
whitenoise==6.7.0